Refactor conversions to separate file, remove logs and comments
