+++ src/types/global.d.ts
@@ -2,14 +2,14 @@
 
 declare global {
   var mongoose: {
-    conn: mongoose // the mongoose connection to use
-    promise: Promise<mongoose> | null // the promise that resolves when the connection is established
-  }
+    conn: mongoose; // the mongoose connection to use
+    promise: Promise<mongoose> | null; // the promise that resolves when the connection is established
+  };
   namespace NodeJS {
     interface ProcessEnv {
-      MONGODB_URI: string // the URI of the mongodb instance to connect to
+      MONGODB_URI: string; // the URI of the mongodb instance to connect to
     }
   }
 }
 
-export {}
+export {};
